requirements.txt
absl-py==0.14.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
google-api-core==2.1.1
google-auth==2.3.0
google-cloud-appengine-logging==0.2.0
google-cloud-audit-log==0.1.1
google-cloud-core==2.1.0
google-cloud-documentai==1.1.0
google-cloud-logging==2.6.0
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
idna==3.3
img2pdf==0.4.2
libcst==0.3.21
lxml==4.6.3
mypy-extensions==0.4.3
packaging==21.0
pikepdf==3.2.0
Pillow==8.4.0
proto-plus==1.19.5
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-bidi==0.4.2
PyYAML==6.0
reportlab==3.6.2
requests==2.26.0
rsa==4.7.2
file:third_party/libseccomp/seccomp-2.5.2-cp39-cp39-linux_x86_64.whl
six==1.16.0
supervisor==4.2.2
tornado==6.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
